bugs fixed

+++ app/src/main/java/com/app/ecarepro/model/LibraryTransactionX.kt
@@ -2,6 +2,6 @@ package com.app.ecarepro.model
 
 data class LibraryTransactionX(
     val bookName: String,
-    val fineAmount: Int,
+    val fineAmount: Double,
     val status: String
 )
